Clean up & refactor code for multiple widgets (WIP, native side only ATM)
